Added a simple batch renderer, will modify it to support text rendering

+++ Source/Graphics/Buffers/IndexBuffer.cpp
@@ -12,7 +12,7 @@ namespace Engine {
 		IndexBuffer::IndexBuffer(const GLushort* data, size_t size) : count(size) {
 			glGenBuffers(1, &id);
 			bind();
-			glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(unsigned int), data, GL_STATIC_DRAW);
+			glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(GLushort), data, GL_STATIC_DRAW);
 			unBind();
 		}
